+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   `@lumx/core`:
+    -   Moved `RawClickable` from `@lumx/react`
+
 ## [4.1.0][] - 2026-01-22
 
 ### Changed